# loop_implementation/api/openapi.yaml
openapi: 3.1.0
info:
  title: Loop Platform API
  version: 0.1.0
  description: |
    Public REST API for the Loop agent platform.

    All endpoints require a Bearer token scoped to a workspace.
    Streaming endpoints additionally support Server-Sent Events (SSE) and WebSocket.
  contact:
    name: Loop Engineering
    email: eng@loop.example
  license:
    name: Apache-2.0
servers:
  - url: https://api.loop.example/v1
    description: Production (na-east — abstract region; concrete cloud may be AWS/Azure/GCP/Alibaba/self-host)
  - url: https://api-eu.loop.example/v1
    description: Production (eu-west — same)
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - BearerAuth: []
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces the caller is a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
...
  /workspaces/{workspace_id}/voice/config:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [Voice]
      summary: Get workspace voice channel configuration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VoiceConfig" }
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags: [Voice]
      summary: Update voice ASR/TTS provider selections
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoiceConfigPatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VoiceConfig" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /workspaces/{workspace_id}/voice/stage:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [Voice]
      summary: Get composed Voice Stage model for a workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VoiceStageModel" }
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /agents:
    post:
      tags: [Agents]
      summary: Create an agent
      parameters:
        - { $ref: "#/components/parameters/ActiveWorkspaceId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
    get:
      tags: [Agents]
      summary: List agents in the active workspace
      parameters:
        - { $ref: "#/components/parameters/ActiveWorkspaceId" }
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              schema: { type: integer }
            X-RateLimit-Remaining:
              schema: { type: integer }
            X-RateLimit-Reset:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /agents/{agent_id}/memory:
    parameters: [{ $ref: "#/components/parameters/AgentId" }]
    get:
      tags: [Memory]
      summary: List runtime memory visible in Memory Studio
      parameters:
        - in: query
          name: user_id
          schema: { type: string }
          required: false
        - in: query
          name: conversation_id
          schema: { type: string, format: uuid }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryStudioEntryList" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /agents/{agent_id}/memory/user/{memory_key}:
    parameters:
      - { $ref: "#/components/parameters/AgentId" }
      - { $ref: "#/components/parameters/MemoryKey" }
    delete:
      tags: [Memory]
      summary: Delete one durable user memory entry with audit trail
      parameters:
        - in: query
          name: user_id
          schema: { type: string }
          required: false
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, description: "Seconds to wait before retrying" }
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer, description: "Unix timestamp of reset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  parameters:
    WorkspaceId:
      in: path
      name: workspace_id
      required: true
      schema: { type: string, format: uuid }
    ActiveWorkspaceId:
      in: header
      name: X-Loop-Workspace-Id
      required: true
      schema: { type: string, format: uuid }
      description: Workspace selected for active agent and audit routes.
...
    AgentId:
      in: path
      name: agent_id
      required: true
      schema: { type: string, format: uuid }
...
    MemoryKey:
      in: path
      name: memory_key
      required: true
      schema: { type: string }
...
  schemas:
...
    ProblemDetail:
      description: RFC 9457 Problem Details for HTTP APIs
      type: object
      required: [type, status, title]
      properties:
        type: { type: string, format: uri, example: "https://api.loop.example/errors/rate-limit" }
        status: { type: integer, example: 429 }
        title: { type: string, example: "Too Many Requests" }
        detail: { type: string }
        instance: { type: string, format: uri }
        request_id: { type: string, description: "Correlation ID for logs" }
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [hobby, pro, team, enterprise] }
        region: { $ref: "#/components/schemas/RegionName" }
        created_at: { type: string, format: date-time }
        role: { type: string, enum: [owner, admin, member, viewer] }
    WorkspaceList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    WorkspaceCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, pattern: "^[a-z0-9][a-z0-9-]*$" }
        region:
          allOf: [{ $ref: "#/components/schemas/RegionName" }]
          default: na-east
          description: Immutable workspace region pin selected at creation.
...
    RegionName:
      type: string
      enum: [na-east, eu-west]
      description: Abstract Loop region from infra/terraform/regions.yaml.
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        active_version: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    AgentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Agent" }
    AgentCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string, pattern: "^[a-z0-9-]+$" }
        description: { type: string }
...
    MemoryStudioEntry:
      type: object
      required:
        - id
        - workspace_id
        - agent_id
        - scope
        - key
        - before
        - after
        - source
        - source_trace
        - retention_policy
        - updated_at
        - writer_version
        - confidence
        - safety_flags
        - deletion_state
        - deletion_reason
        - replay_impact
      properties:
        id: { type: string }
        workspace_id: { type: string, format: uuid }
        agent_id: { type: string, format: uuid }
        scope: { type: string, enum: [user, bot, session] }
        user_id: { type: string, nullable: true }
        key: { type: string }
        before: { type: string }
        after: { type: string }
        source: { type: string }
        source_trace: { type: string }
        retention_policy: { type: string }
        updated_at: { type: string }
        writer_version: { type: string }
        confidence: { type: string, enum: [high, medium, low, unsupported] }
        safety_flags:
          type: array
          items:
            type: string
            enum: [none, pii, secret-like, conflict, stale, weak-evidence]
        deletion_state: { type: string, enum: [available, blocked, queued] }
        deletion_reason: { type: string }
        replay_impact: { type: string }
    MemoryStudioEntryList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/MemoryStudioEntry" }
    VoiceNumber:
      type: object
      required: [id, e164, label, region, provisioned_at_ms]
      properties:
        id: { type: string }
        e164: { type: string }
        label: { type: string }
        region: { type: string }
        provisioned_at_ms: { type: integer }
    VoiceConfig:
      type: object
      required: [workspace_id, numbers, asr_provider, tts_provider]
      properties:
        workspace_id: { type: string, format: uuid }
        numbers:
          type: array
          items: { $ref: "#/components/schemas/VoiceNumber" }
        asr_provider: { type: string, enum: [deepgram, whisper, google] }
        tts_provider: { type: string, enum: [elevenlabs, openai, polly] }
    VoiceConfigPatch:
      type: object
      properties:
        asr_provider: { type: string, enum: [deepgram, whisper, google] }
        tts_provider: { type: string, enum: [elevenlabs, openai, polly] }
    VoiceTranscriptTurn:
      type: object
      required: [id, speaker, text, timestamp]
      properties:
        id: { type: string }
        speaker: { type: string, enum: [caller, agent, tool] }
        text: { type: string }
        timestamp: { type: string }
    VoiceLatencySpan:
      type: object
      required: [id, label, ms, budgetMs, status]
      properties:
        id: { type: string }
        label: { type: string }
        ms: { type: integer }
        budgetMs: { type: integer }
        status: { type: string, enum: [ok, watch, over] }
    VoiceStageConfig:
      type: object
      required: [asr, tts, bargeIn, voice, phoneNumber]
      properties:
        asr: { type: string }
        tts: { type: string }
        bargeIn: { type: boolean }
        voice: { type: string }
        phoneNumber: { type: string }
    VoiceEvalCase:
      type: object
      required: [id, name, passRate, coverage, evidenceRef]
      properties:
        id: { type: string }
        name: { type: string }
        passRate: { type: integer }
        coverage: { type: string }
        evidenceRef: { type: string }
    VoiceDemoLink:
      type: object
      required: [id, label, expiresIn, scope, audited]
      properties:
        id: { type: string }
        label: { type: string }
        expiresIn: { type: string }
        scope: { type: string }
        audited: { type: boolean }
    VoiceStageModel:
      type: object
      required:
        - agentName
        - callState
        - queuedSpeech
        - transcript
        - waveform
        - spans
        - config
        - evals
        - demoLinks
      properties:
        agentName: { type: string }
        callState: { type: string, enum: [dev, staging, production] }
        queuedSpeech: { type: string }
        transcript:
          type: array
          items: { $ref: "#/components/schemas/VoiceTranscriptTurn" }
        waveform:
          type: array
          items: { type: integer }
        spans:
          type: array
          items: { $ref: "#/components/schemas/VoiceLatencySpan" }
        config: { $ref: "#/components/schemas/VoiceStageConfig" }
        evals:
          type: array
          items: { $ref: "#/components/schemas/VoiceEvalCase" }
        demoLinks:
          type: array
          items: { $ref: "#/components/schemas/VoiceDemoLink" }
...